requirements.txt
fastapi
uvicorn
numpy<2
pillow
onnxruntime
qdrant-client
python-multipart
requests
